Create support module to DRY up common setup
This created a support module that we can use to extend RSpec and allow
use to through common setup and test items in so we can DRY up the
tests.

For example this allows us to remove the setup block for the TmdbRexx
client that has to happen on every test for the client to actually do
anything.

This could be further refined later or split apart to accomplish
different tasks at a more granualar level.