Remove hashie from gempspec and gem
Hashie as it turns out was a not so great idea. It ends up polluting the
interface with a lot of unnecessary stuff and really does not give much
over methood access instead of hash access for our uses.

This will in turn cause the responses to be returned in JSON which is
how it really should be since we are just wrapping an api.